Rename presidio_data_generator.py to presidio_sentence_faker.py
